ch07/pyproject.toml
[project]
name = "ch07"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
    "tabulate>=0.9.0",
    "tzdata>=2025.2",
    "yfinance>=0.2.65",
]
